cleaned file directories
