+++ RamenShopApp/View/Common/Constants.swift
@@ -10,6 +10,7 @@ import Foundation
 class Constants {
     static let ADMIN_EMAIL = "korokorokoro.nn99@gmail.com"
     static let EMPTY = ""
+    static let OK = "OK"
     
     static let NO_NAME = "no name"
     static let REQUIRE_COMMENT = "enter comment"
@@ -18,4 +19,14 @@ class Constants {
     static let REQUEST_STATUS_MESSAGE_OTHER = "※If you wanna create new request, you need to click the above button."
     
     static let REQUIRED_ZOOM_LIMIT = 19.0
+    
+    static let SIGNUP_UPPERCASE = "Sign Up"
+    static let SIGNUP = "sign up"
+    static let EMAIL = "email"
+    static let PASSWORD = "passsword"
+    
+    static let SENT_EMAIL_TITLE = "Sent Email!"
+    static let SENT_EMAIL_MESSAGE = "We sent Email to your adress, so please check it."
+    
+    static let SIGNUP_ERROR_TITLE = "Signup Error"
 }
